Finished Most of the ui
Preplanning and pseudocode helped start quite strong.
All global variable are set to the top of the js to
stop repetitions. I attempted to use sectorAll but
was unable to really make is work. So it is a
little more cluttered then I would like but
it is over all working.